Refactor Check All to component
