+++ components/form-builder/app/responses/DownloadTable.tsx
@@ -26,11 +26,7 @@ import {
 import { getDaysPassed } from "@lib/clientHelpers";
 import { Alert } from "@components/globals";
 import { logMessage } from "@lib/logger";
-import {
-  CheckAllIcon,
-  CheckBoxEmptyIcon,
-  CheckIndeterminateIcon,
-} from "@components/form-builder/icons";
+import { CheckAll } from "./CheckAll";
 
 interface DownloadTableProps {
   vaultSubmissions: VaultSubmissionList[];
@@ -69,57 +65,6 @@ export const DownloadTable = ({
   });
 
   const MAX_FILE_DOWNLOADS = responseDownloadLimit;
-
-  enum allCheckedState {
-    NONE = "NONE",
-    SOME = "SOME",
-    ALL = "ALL",
-  }
-
-  /**
-   * Are all items checked, some items checked, or no items checked?
-   * @returns {allCheckedState} The current state of the checked items.
-   */
-  const checkAllStatus = () => {
-    if (tableItems.checkedItems.size === 0) {
-      return allCheckedState.NONE;
-    }
-    if (tableItems.checkedItems.size === tableItems.sortedItems.length) {
-      return allCheckedState.ALL;
-    }
-    return allCheckedState.SOME;
-  };
-
-  /**
-   * Check all items if none are checked, otherwise uncheck all items.
-   */
-  const handleCheckAll = () => {
-    if (checkAllStatus() === allCheckedState.NONE) {
-      vaultSubmissions.forEach((submission) => {
-        tableItemsDispatch({
-          type: TableActions.UPDATE,
-          payload: { item: { name: submission.name, checked: true } },
-        });
-      });
-    } else {
-      tableItems.checkedItems.forEach((_, key) => {
-        tableItemsDispatch({
-          type: TableActions.UPDATE,
-          payload: { item: { name: key, checked: false } },
-        });
-      });
-    }
-
-    // Show or hide errors depending
-    if (tableItems.checkedItems.size > MAX_FILE_DOWNLOADS && !errors.maxItemsError) {
-      setErrors({ ...errors, maxItemsError: true });
-    } else if (errors.maxItemsError) {
-      setErrors({ ...errors, maxItemsError: false });
-    }
-    if (tableItems.checkedItems.size > 0 && errors.noItemsError) {
-      setErrors({ ...errors, noItemsError: false });
-    }
-  };
 
   const handleChecked = (e: React.ChangeEvent<HTMLInputElement>) => {
     const name = e.target.id;
@@ -276,43 +221,13 @@ export const DownloadTable = ({
         <thead className="border-b-2 border-[#6a6d7b]">
           <tr>
             <th className="p-4 text-center">
-              <span
-                className="cursor-pointer"
-                role="checkbox"
-                aria-checked={
-                  checkAllStatus() === allCheckedState.ALL
-                    ? "true"
-                    : checkAllStatus() === allCheckedState.SOME
-                    ? "mixed"
-                    : "false"
-                }
-                tabIndex={0}
-                onClick={handleCheckAll}
-                onKeyDown={(e) => {
-                  if (e.key === "Enter" || e.key === " ") {
-                    handleCheckAll();
-                  }
-                }}
-              >
-                {checkAllStatus() === allCheckedState.ALL && (
-                  <CheckAllIcon
-                    title={t("downloadResponsesTable.header.deselectAll")}
-                    className="m-auto h-6 w-6"
-                  />
-                )}
-                {checkAllStatus() === allCheckedState.SOME && (
-                  <CheckIndeterminateIcon
-                    title={t("downloadResponsesTable.header.deselectAll")}
-                    className="m-auto h-6 w-6"
-                  />
-                )}
-                {checkAllStatus() === allCheckedState.NONE && (
-                  <CheckBoxEmptyIcon
-                    title={t("downloadResponsesTable.header.selectAll")}
-                    className="m-auto h-6 w-6"
-                  />
-                )}
-              </span>
+              <CheckAll
+                responseDownloadLimit={responseDownloadLimit}
+                tableItems={tableItems}
+                tableItemsDispatch={tableItemsDispatch}
+                errors={errors}
+                setErrors={setErrors}
+              />
             </th>
             <th className="p-4 text-left">{t("downloadResponsesTable.header.number")}</th>
             <th className="p-4 text-left">{t("downloadResponsesTable.header.status")}</th>
